+++ source/commands/club-leave.js
@@ -2,6 +2,7 @@ const { MessageFlags, InteractionContextType } = require('discord.js');
 const { CommandWrapper } = require('../classes');
 const { updateClub, updateListReference, getClub } = require('../engines/referenceEngine.js');
 const { commandMention } = require('../util/textUtil.js');
+const { createClubRecruitmentEvent } = require('../engines/clubEngine.js');
 
 const mainId = "club-leave";
 module.exports = new CommandWrapper(mainId, "Leave this club", null, [InteractionContextType.Guild], 3000,
@@ -19,10 +20,11 @@ module.exports = new CommandWrapper(mainId, "Leave this club", null, [Interactio
 					.catch(console.error);
 				interaction.guild.channels.resolve(club.voiceChannelId).permissionOverwrites.delete(interaction.user, `/${mainId}`)
 					.catch(console.error);
-				updateListReference(interaction.guild.channels, "club");
 				updateClub(club);
 				interaction.reply({ content: `${interaction.user} has left this club.`, flags: MessageFlags.SuppressNotifications })
 					.catch(console.error);
+				updateListReference(interaction.guild.channels, "club");
+				createClubRecruitmentEvent(club, interaction.guild);
 			}
 		} else {
 			interaction.reply(`Please use the \`/${mainId}\` command from the club's text channel.`)